1. add sign (direction) in gd.rs; 2. make x in test1.cpp smaller to avoid overflow

+++ AD-tests/test1/test1.cpp
@@ -29,7 +29,7 @@ int main (int argc, char** argv) {
     }
 
     int len = 1;
-    int16_t _x = 0;
+    int8_t _x = 0;
     ret = fread(&_x, sizeof _x, 1, fp);
     Int x = Int(_x, 1);
     x.get_dx(-1);
